Use new slack_sdk instead of slackclient

+++ bellchan/lib/heroku/heroku.py
@@ -19,4 +19,7 @@ class Heroku:
 
     @classmethod
     def restart(cls):
+        if not Settings.HEROKU_API_KEY:
+            return
+
         return requests.delete(cls.URLS['restart'], headers=cls.get_basic_headers())
